Use express workflow for repo sub-objects fan out
Step Functions Standard Workflows cost $0.025 / 1000 transitions, which is actual money for these crawlers when
dealing with 100s or 1000s of repos. So switch to an Express
Workflow when crawling child elements of a repo.

Doing this meant introducing a new Step Functions invocation,
passing a list of repositories to it. The repository object
I was using before was quite large, and causing problems,
so just use repository summary objects instead when concerned
with passing data between different crawlers.