Introduce Github Workflow as specific entity / object
We can't get Github webhooks for Workflows, but we can read them from the API.
This is useful since for some workflows, like Dependabot, the name on a workflow
run is NOT the same as the name of the workflow itself.

So now - crawl all workflows during installation crawl, and then specifically
lookup a workflow during webhook processing if it isn't already in the database.

Had to change the remote test to use pushes instead of workflow runs since
otherwise the remote test would try to lookup the workflow.